requirements.txt
flask==3.1.3
linkedin-api==2.3.1
python-dotenv==1.0.1
